Updated the deployment pipeline and page styling
1. Page styling was rendering the site unusable on mobile devices and
other width constricted viewports, to alleviate that, a proper mobile
view bar was included.

2. Due to the duplication of layouts for both mobile and standard view, a
bug was introduced in which html ID's were duplicated. This meant that
anchor tags did not work, rendering the table of contents useless. Since
mobile contains no ToC, extra functionality was introduced to strip ID
tags from the mobile content. This is a messy fix, but was chosen to
allow for the continued use of pandoc to convert files, without too
much manipulation of the generated html page.

3. Deployment scripts were updated to be more resilient and present
better failure states. The update and deploy functions have been
broken into two scripts. Very, very crude uptime monitoring occurs
at the end of the script by attempting to curl the newly deployed
site and adjusting the "success" message in the event of any non-zero
exit code.